+++ src/libs/BaseComplexDataType/interfaces/IBaseComplexDataType.ts
@@ -16,3 +16,18 @@ export interface IBaseComplexDataType_ {
      */
     isInstanceOf<T>(this: new (...args: never[]) => T, obj: unknown): obj is T;
 }
+
+/**
+ * Checks if the object is an {@link IBaseComplexDataType}.
+ * @param obj The object to check.
+ * @returns Whether the object is an {@link IBaseComplexDataType}.
+ */
+export function isIBaseComplexDataType(
+    obj: unknown,
+): obj is IBaseComplexDataType {
+    return (
+        obj != null &&
+        typeof obj === 'object' &&
+        IBaseComplexDataTypeSymbol in (obj as object)
+    );
+}
